+++ library/src/main/java/com/paypal/messages/PayPalMessageView.kt
@@ -322,8 +322,12 @@ class PayPalMessageView @JvmOverloads constructor(
 	override fun onDetachedFromWindow() {
 		super.onDetachedFromWindow()
 		// The modal will not dismiss (destroy) itself, it will only hide/show when opening and closing
-		// so we need to cleanup the modal instance if the message is removed
-		this.modal?.dismiss()
+		// so we need to clean up the modal instance if the message is removed
+		this.modal?.let {
+			if (it.isAdded) {
+				it.dismiss()
+			}
+		}
 	}
 
 	/**
